add sendEmail
