+++ helpers/asyncWrapper.js
@@ -1,0 +1,7 @@
+const asyncWrapper = (controller) => {
+  return (req, res, next) => {
+    controller(req, res).catch(next);
+  };
+};
+
+module.exports = asyncWrapper;
