+++ helpers/errorHandler.js
@@ -1,11 +1,5 @@
 const { checkError } = require("./errors");
 
-const asyncWrapper = (controller) => {
-  return (req, res, next) => {
-    controller(req, res).catch(next);
-  };
-};
-
 const errorHandler = (error, req, res, next) => {
   if (error instanceof checkError) {
     return res.status(error.status).json({ message: error.message });
@@ -14,7 +8,4 @@ const errorHandler = (error, req, res, next) => {
   res.status(500).json({ message: error.message });
 };
 
-module.exports = {
-  asyncWrapper,
-  errorHandler,
-};
+module.exports = errorHandler;
